requirements.txt
mysql-connector-python
mypy
pytest
